attaching separate video and image forward functions
