+++ bitmind/base/miner.py
@@ -59,7 +59,11 @@ class BaseMinerNeuron(BaseNeuron):
         # Attach determiners which functions are called when servicing a request.
         bt.logging.info(f"Attaching forward function to miner axon.")
         self.axon.attach(
-            forward_fn=self.forward,
+            forward_fn=self.forward_image,
+            blacklist_fn=self.blacklist,
+            priority_fn=self.priority,
+        ).attach(
+            forward_fn=self.forward_video,
             blacklist_fn=self.blacklist,
             priority_fn=self.priority,
         )
